changed experience datatype

+++ backend/user/models.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import AbstractUser
 
 class User(AbstractUser):
     institution = models.TextField(editable=True, null=False)
-    experience = models.IntegerField(editable=True, null=False)
+    experience = models.PositiveSmallIntegerField(editable=True, null=False)
     approved = models.BooleanField(default=False)
 
     REQUIRED_FIELDS = ["first_name", "last_name", "email", "institution", "experience"]
